Updated reimbursement policy on submission timeline
Updated the policy to take out the line that reimbursements can be denied if past the 60 day limit. 

Per CI attorney, the standard is that "unless an employee has waited more than 4 years (the statue of limitations for "indemnification" from an employer) reimbursements should not be denied based on time of submission.