clarify Business Admin should receive requests
Lex pointed out that it is unclear here to the community who to send the documentation (receipts, etc...) to in order to start the reimbursement process. He guessed me because he is more aware of staff rules, but I think clarifying here is a good idea. These additions reflect the process that has been and is currently followed for community reimbursement.