Replaced TBA with "Do not use"
Replaced TBA (which could mean "To Be Avoided" or "To Be Announced", the latter potentially meaning that students could use this material for demos) with "Do not use", which clearly indicates that this material should not be used for training demos. 

Ideally, we could then point to this page if students do try to choose one of the lessons to show that these materials cannot be used?