Add Loops in R to what avoid
I added loops in R to what to avoid. I accidentally picked this today for a checkout as it was not obvious to me that this actually supplementary material and dependent on an to avoid section "Analyzing Multiple Data Sets". Thought making this clearly visible will avoid issues in future.