Added *recommended* committee guidelines
Software Carpentry had this guidance for committees, but when we merged into The Carpentries, we restructured to create both committees and Task Forces, and wanted to give some time to see how things came together before creating guidelines. We now have a workflow and Task Force Charter model, and want to work with the committees to establish guidelines and consistent documentation and structure for committees.