Add language about use of video
We recently received feedback that instructions should warn trainees that they will be asked to turn their video on. This likely results from an increase in online events, where many schools have made it policy that students cannot be required to have their cameras on. This PR adds some suggested language -- not sure if this should be circulated for discussion before adding?